+++ static/admin/assets/js/hash.js
@@ -1,0 +1,9 @@
+export function simpleHash(str) {
+    let hash = 729790;
+
+    for (let i = 0; i < str.length; i++) {
+        hash = ((hash << 5) + hash) + str.charCodeAt(i);
+    }
+
+    return (hash >>> 0).toString(16).slice(0, 8); // Convert to positive hex
+}